HCS-4: add couple spending and couple spending category policy

+++ app/Http/Livewire/CoupleSpendings/Destroy.php
@@ -4,14 +4,18 @@ namespace App\Http\Livewire\CoupleSpendings;
 
 use App\Models\CoupleSpending;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Destroy extends Component
 {
+    use AuthorizesRequests;
+
     public ?CoupleSpending $coupleSpending = null;
 
     public function save(): void
     {
+        $this->authorize('delete', $this->coupleSpending);
 
         $this->coupleSpending->delete();
